example page for wildcards
Made an example so that wildcards work off the bat. loads a list of at least four urls the the octomatization code finds the urls that contain the wildcards and takes them out saving memory.

Also, added code so that if size of wildcard is 0 it skips octomization saving user time.