Update Blockade.kt
addresses:

unused variable.

There can't be any ',' left in the string because you are using split(","). Additionally you can use trim() instead of removePrefix(" ") and removeSuffix(" ") which automatically removes all spaces at the beginning and end of the string, making it more secure.

This if is unnecessary. Simply remove it and move Log.d() and return ruleset behind the end of the for-loop.