Update Pages.kt
 ps100000 5 days ago Contributor

Don't replace the official API.
@Thomas499
Thomas499 5 days ago Author Contributor

didn't know how else to get code to work so that you could compile it and wildcard was enabled (out of box). Without this, it would take a while to figure out how to implement wildcards.
@ps100000
ps100000 5 days ago Contributor

Just add the information to your PR or add a link to a second branch where you have a "ready to try"-version of the code.