Changing memory related values
Based on a list with 1.327.728 unique hostnames, the average hostname lenght is 20,41 char long. As fractions aren't supported, it's better to round up to 21.
Unicode characters use 2 bytes.
I ran a test for a few days to see how much RAM one hostname consumes in average, it is around 130 bytes