android: change detection order of wifi and mobile networks
Seems like some networks can have both Transports: Cellural and Wifi. Since we
prioritise Wifi networks if any is active, we seem to run into a bug with some
specific mobile networks. Changing detection priority here, to hopefully
solve it, in a way that should not have side effects. As usual with Android,
it's just a guess and hopefully will work (: