ios: add timeout to netx state request and messages
It seems something is sometimes hanging indefinitely when asking NETX for
its current state. From my tests it doesn't seem to be the messaging,
more likely a problem getting the Manager object from the system api.
I added a 10s timeout for NETX messages, since there should be a timeout
anyway (as http requests are made through it when vpn is on, for example),
and also 5s timeout getting the Manager object. Otherwise error will
be returned and the standard error handling will fire, instead of hanging
on the "in progress" state as we experienced. Real life testing tbd.