ios: prevent mocked keys in non-tests runs
Mocked keys were persisted out of the test run and then attempted
to be used in the actual builds. This prevents it, but ideally
tests should be probably isolated from the outside env (persistence,
network). For now they are not.

Also added valid keys as mocked set, so that Mocked builds can actually
call the API and are as similar to actual production code as possible.