android: relax privetdns warning for libre
It's too aggressive assuming no Private DNS is set. Changed so that its
ok if users correct private DNS is set, as some users will use Cloud
together with libre (v5) app. It's a corner case. One that we experienced
irl is that user wants to use Cloud while not using Google Play Services,
so he cannot install v6 from the store.

The side effect of this change may be that user won't be notified
about having the Private DNS set when using v5, and this may lead
to conn issues possibly when Plus is activated (not sure).