add a feature to save crash log
A command to save a separate log on crash. This cannot be
hooked up on ios, but on android yes.

On app start, will check for this log and propose to share it.
Will delete the log file
Changed log filenames for consistency
Changed how overlays are handled for consistency