delete obsolete test
The idea is that the common lib is where the tests are and
that lib is well tested. The android/ios codebases do not
have any tests so far, but they can be added as CI is ready
to run them. Maybe UI integration tests will land there
one day. Probably not:)