remove url_launcher dep (apple reject)
So because we have the old project it is setup in a way that is suboptimal
and I believe it pulls the flutter plugin without proper signing, what
makes apple reject it.

The easiest way I can think to fix this without redoing the whole
project setup and us pods (which I'm not sure will fix the problem),
is to try to implement the missing url launcher platform code
that is expected to be provided on ios / android side.