v6: fix gateway handling related to autostart
It was stil possible to autostart kick in before
the gateways are fetched, resulting in comparing
to empty array, and in an error when starting
VPN because gateway was not found.

Made use of Value to store gateways so that
it will wait until they are loaded.

Adedd a small delay for the autostart job
(after making it execute async before),
so that it will trigger later in the startup
flow. The async execution already made sure
it won't deadlock.